break out and structure the code
