+++ models/Thought.js
@@ -1,0 +1,16 @@
+import mongoose from "mongoose"
+
+const thoughtSchema = new mongoose.Schema({
+
+  message: String,
+  hearts: {
+    type: Number,
+    default: 0
+  },
+  createdAt: {
+    type: Date,
+    default: Date.now
+  }
+})
+
+export const Thought = mongoose.model("Thought", thoughtSchema)
